requirements.txt
streamlit
google-generativeai
faiss-cpu
pypdf
sentence-transformers
numpy